+++ src/main/java/ch/hsr/graph/DepthFirstSearch.java
@@ -7,7 +7,6 @@ import ch.hsr.adv.lib.core.logic.util.ADVException;
 import ch.hsr.adv.lib.graph.logic.GraphModule;
 import ch.hsr.adv.lib.graph.logic.domain.ADVVertex;
 import ch.hsr.adv.lib.stack.logic.StackModule;
-import ch.hsr.adv.lib.stack.logic.domain.EmptyStackException;
 import ch.hsr.graph.model.Edge;
 import ch.hsr.graph.model.Graph;
 import ch.hsr.graph.model.Vertex;
